requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
nats-py
pydantic
python-dotenv
jinja2
httpx
pytest
pytest-asyncio
python-multipart
